+++ exo.jcr.component.core/src/main/java/org/exoplatform/services/jcr/impl/clean/rdbms/scripts/MySQLCleaningScipts.java
@@ -147,8 +147,11 @@ public class MySQLCleaningScipts extends DBCleaningScripts
       scripts.add("ALTER TABLE " + valueTableName + " RENAME TO " + valueTableName + "_OLD");
       scripts.add("ALTER TABLE " + itemTableName + " RENAME TO " + itemTableName + "_OLD");
       scripts.add("ALTER TABLE " + refTableName + " RENAME TO " + refTableName + "_OLD");
-      scripts.add("ALTER TABLE "+itemTableName+"_SEQ RENAME TO "+itemTableName+"_SEQ_OLD");
-      scripts.add("DROP FUNCTION " + itemTableName + "_NEXT_VAL");
+      if (useSequence)
+      {
+         scripts.add("ALTER TABLE " + itemTableName + "_SEQ RENAME TO " + itemTableName + "_SEQ_OLD");
+         scripts.add("DROP FUNCTION " + itemTableName + "_NEXT_VAL");
+      }
 
       return scripts;
    }
@@ -163,18 +166,22 @@ public class MySQLCleaningScipts extends DBCleaningScripts
       scripts.add("ALTER TABLE " + itemTableName + "_OLD RENAME TO " + itemTableName);
       scripts.add("ALTER TABLE " + valueTableName + "_OLD RENAME TO " + valueTableName);
       scripts.add("ALTER TABLE " + refTableName + "_OLD RENAME TO " + refTableName);
-      scripts.add("ALTER TABLE "+itemTableName+"_SEQ_OLD RENAME TO "+itemTableName+"_SEQ");
-      try
+      if (useSequence)
       {
-         scripts.add(DBInitializerHelper.getObjectScript("CREATE FUNCTION " + itemTableName + "_NEXT_VAL", multiDb, dialect, wsEntry));
-      }
-      catch (RepositoryConfigurationException e)
-      {
-         throw new DBCleanException(e);
-      }
-      catch (IOException e)
-      {
-         throw new DBCleanException(e);
+         scripts.add("ALTER TABLE " + itemTableName + "_SEQ_OLD RENAME TO " + itemTableName + "_SEQ");
+         try
+         {
+            scripts.add(DBInitializerHelper.getObjectScript("CREATE FUNCTION " + itemTableName + "_NEXT_VAL", multiDb, dialect, wsEntry));
+         }
+         catch (RepositoryConfigurationException e)
+         {
+
+            throw new DBCleanException(e);
+         }
+         catch (IOException e)
+         {
+            throw new DBCleanException(e);
+         }
       }
 
       return scripts;
@@ -187,7 +194,10 @@ public class MySQLCleaningScipts extends DBCleaningScripts
    {
       List<String> scripts = new ArrayList<String>();
 
-      scripts.add("DROP TABLE "+itemTableName+"_SEQ_OLD");
+      if (useSequence)
+      {
+         scripts.add("DROP TABLE " + itemTableName + "_SEQ_OLD");
+      }
       scripts.addAll(super.getOldTablesDroppingScripts());
 
       return scripts;
@@ -200,7 +210,10 @@ public class MySQLCleaningScipts extends DBCleaningScripts
    {
       List<String> scripts = new ArrayList<String>();
 
-      scripts.add("DROP TABLE "+itemTableName+"_SEQ");
+      if (useSequence)
+      {
+         scripts.add("DROP TABLE " + itemTableName + "_SEQ");
+      }
       scripts.addAll(super.getTablesDroppingScripts());
 
       return scripts;
